Fixing thin clients.
Moved trinity.windowsEvents to windowsutils.windowsEvents
Those constants are also in eve/devtools/script/uiEventListenerConsts.py. I noticed that all the consts in uiEventListenerConsts are in windowsEvents but it is doing some reverse but there are more in windowsEvents and we have duplicate names for each key so I was not sure how to handle the reverse lookups.

Review: http://codecollaborator/ui#review:id=5068

Binaries submitted in 855360
Build time : 8 minutes 47.851000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 855351]