Took out the matrices that are sent to the shader for the turretsets and replaced them with a translation vector and a rotation quaternion. This is applied for both bones and models
Made the bonecount dynamic so we always draw as many turrets as we can for a turretset.
The maximum boneCount for a turretSet is 74 and the maximum number of turrets is 24.
So if a turret model has 4 bones then we get floor(74/4) = 18 turrets that can be displayed.

Task: http://jira.ccpgames.com/browse/GFX-317

Review: http://codecollaborator/ui#review:id=5732

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 879891]